PythonModel/requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
opencv-python==4.8.1.78
mediapipe==0.10.21
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
imgaug==0.4.0
albumentations==1.3.1
tqdm==4.65.0
pillow==10.0.0
pandas==2.0.3
seaborn==0.12.2
tensorboard==2.15.0 